Reduce complexity of common module by splitting it up
