+++ resources/lib/common/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# pylint: disable=unused-import, wildcard-import
+"""Common plugin operations and utilities"""
+from __future__ import unicode_literals
+
+from .globals import *
+from .logging import *
+from .ipc import *
+from .videoid import *
+from .credentials import *
+from .storage import *
+from .fileops import *
+from .kodiops import *
+from .pathops import *
+from .misc_utils import *
