+++ resources/lib/common/credentials.py
@@ -1,0 +1,110 @@
+# -*- coding: utf-8 -*-
+"""Handling of account credentials"""
+from __future__ import unicode_literals
+
+from .globals import ADDON
+
+
+class MissingCredentialsError(Exception):
+    """There are no stored credentials to load"""
+    pass
+
+__BLOCK_SIZE__ = 32
+__CRYPT_KEY__ = None
+
+
+def __crypt_key():
+    """
+    Lazily generate the crypt key and return it
+    """
+    # pylint: disable=global-statement
+    global __CRYPT_KEY__
+    if not __CRYPT_KEY__:
+        __CRYPT_KEY__ = __uniq_id()
+    return __CRYPT_KEY__
+
+
+def __uniq_id():
+    """
+    Returns a unique id based on the devices MAC address
+    """
+    import uuid
+    mac = uuid.getnode()
+    if (mac >> 40) % 2:
+        from platform import node
+        mac = node()
+    return uuid.uuid5(uuid.NAMESPACE_DNS, str(mac)).bytes
+
+
+def encrypt_credential(raw):
+    """
+    Encodes data
+
+    :param data: Data to be encoded
+    :type data: str
+    :returns:  string -- Encoded data
+    """
+    # pylint: disable=invalid-name,import-error
+    import base64
+    from Cryptodome import Random
+    from Cryptodome.Cipher import AES
+    from Cryptodome.Util import Padding
+    raw = bytes(Padding.pad(data_to_pad=raw, block_size=__BLOCK_SIZE__))
+    iv = Random.new().read(AES.block_size)
+    cipher = AES.new(__crypt_key(), AES.MODE_CBC, iv)
+    return base64.b64encode(iv + cipher.encrypt(raw))
+
+
+def decrypt_credential(enc):
+    """
+    Decodes data
+
+    :param data: Data to be decoded
+    :type data: str
+    :returns:  string -- Decoded data
+    """
+    # pylint: disable=invalid-name,import-error
+    import base64
+    from Cryptodome.Cipher import AES
+    from Cryptodome.Util import Padding
+    enc = base64.b64decode(enc)
+    iv = enc[:AES.block_size]
+    cipher = AES.new(__uniq_id(), AES.MODE_CBC, iv)
+    decoded = Padding.unpad(
+        padded_data=cipher.decrypt(enc[AES.block_size:]),
+        block_size=__BLOCK_SIZE__).decode('utf-8')
+    return decoded
+
+
+def get_credentials():
+    """
+    Retrieve stored account credentials.
+    :return: The stored account credentials or an empty dict if none exist.
+    """
+    email = ADDON.getSetting('email')
+    password = ADDON.getSetting('password')
+    verify_credentials(email, password)
+    try:
+        return {
+            'email': decrypt_credential(email),
+            'password': decrypt_credential(password)
+        }
+    except ValueError:
+        raise MissingCredentialsError(
+            'Existing credentials could not be decrypted')
+
+
+def set_credentials(email, password):
+    """
+    Encrypt account credentials and save them to the settings.
+    Does nothing if either email or password are not supplied.
+    """
+    if email and password:
+        ADDON.setSetting('email', encrypt_credential(email))
+        ADDON.setSetting('password', encrypt_credential(password))
+
+
+def verify_credentials(email, password):
+    """Verify credentials for plausibility"""
+    if not email or not password:
+        raise MissingCredentialsError()
