+++ resources/lib/kodi/library.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 
 import os
 import codecs
+from datetime import datetime, timedelta
 
 import xbmc
 
@@ -76,28 +77,45 @@ def is_in_library(videoid):
 
 
 def compile_tasks(videoid):
-    """Compile a list of tasks for items based on the videoid.
-    If videoid represents a show or season, tasks will be generated for
-    all contained seasons and episodes"""
+    """Compile a list of tasks for items based on the videoid"""
     metadata = api.metadata(videoid)
     if videoid.mediatype == common.VideoId.MOVIE:
-        name = '{title} ({year})'.format(
-            title=metadata['title'],
-            year=metadata['year'])
-        return [_create_item_task(name, FOLDER_MOVIES, videoid, name, name)]
-    elif videoid.mediatype == common.VideoId.SHOW:
-        return [
-            task for season in metadata['seasons']
-            for task in _compile_season_tasks(
-                videoid.derive_season(season['id']), metadata, season)]
+        return _create_movie_task(videoid, metadata)
+    elif videoid.mediatype in common.VideoId.TV_TYPES:
+        return _create_tv_tasks(videoid, metadata)
+
+    raise ValueError('Cannot handle {}'.format(videoid))
+
+
+def _create_movie_task(videoid, metadata):
+    """Create a task for a movie"""
+    name = '{title} ({year})'.format(
+        title=metadata['title'],
+        year=metadata['year'])
+    return [_create_item_task(name, FOLDER_MOVIES, videoid, name, name)]
+
+
+def _create_tv_tasks(videoid, metadata):
+    """Create tasks for a show, season or episode.
+    If videoid represents a show or season, tasks will be generated for
+    all contained seasons and episodes"""
+    if videoid.mediatype == common.VideoId.SHOW:
+        return _compile_show_tasks(videoid, metadata)
     elif videoid.mediatype == common.VideoId.SEASON:
         return _compile_season_tasks(
             videoid, metadata, common.find_season(videoid.seasonid,
                                                   metadata))
-    elif videoid.mediatype == common.VideoId.EPISODE:
-        return [_create_episode_task(videoid, metadata)]
+    return [_create_episode_task(videoid, metadata)]
 
-    raise ValueError('Cannot handle {}'.format(videoid))
+
+def _compile_show_tasks(videoid, metadata):
+    """Compile a list of task items for all episodes of all seasons
+    of a tvshow"""
+    # This nested comprehension is nasty but neccessary. It flattens
+    # the task lists for each season into one list
+    return [task for season in metadata['seasons']
+            for task in _compile_season_tasks(
+                videoid.derive_season(season['id']), metadata, season)]
 
 
 def _compile_season_tasks(videoid, metadata, season):
@@ -137,12 +155,22 @@ def export_item(item_task, library_home):
         library_home, item_task['section'], item_task['destination'])
     export_filename = os.path.join(
         destination_folder, item_task['filename'] + '.strm')
+    _create_destination_folder(destination_folder)
+    _write_strm_file(item_task, export_filename)
+    _add_to_library(item_task['videoid'], export_filename)
+
+
+def _create_destination_folder(destination_folder):
+    """Create destination folder, ignore error if it already exists"""
     try:
         os.makedirs(xbmc.translatePath(destination_folder))
     except OSError as exc:
         if exc.errno != os.errno.EEXIST:
             raise
 
+
+def _write_strm_file(item_task, export_filename):
+    """Write the playable URL to a strm file"""
     try:
         with codecs.open(xbmc.translatePath(export_filename),
                          mode='w',
@@ -157,7 +185,6 @@ def export_item(item_task, library_home):
                         .format(export_filename))
         else:
             raise
-    _add_to_library(item_task['videoid'], export_filename)
 
 
 def _add_to_library(videoid, export_filename):
@@ -188,3 +215,28 @@ def update_item(item_task, library_home):
     """Remove and then re-export an item to the Kodi library"""
     remove_item(item_task)
     export_item(item_task, library_home)
+
+
+def _update_running():
+    update = common.ADDON.getSetting('update_running') or None
+    if update:
+        starttime = common.strp(update, '%Y-%m-%d %H:%M')
+        if (starttime + timedelta(hours=6)) <= datetime.now():
+            common.ADDON.setSetting('update_running', 'false')
+            common.warn('Canceling previous library update: duration >6 hours')
+        else:
+            common.debug('DB Update already running')
+            return True
+    return False
+
+
+def update_library():
+    """
+    Update the local Kodi library with new episodes of exported shows
+    """
+    if not _update_running():
+        common.info('Triggering library update')
+        xbmc.executebuiltin(
+            ('XBMC.RunPlugin(plugin://{}/?action=export-new-episodes'
+             '&inbackground=True)')
+            .format(common.ADDON_ID))
